Rename and remove mailhog references
